refactor: flatten settings panels folder
The panels/ subfolder grouped four *Settings files that are already
grouped by their shared suffix. Flatten to reduce nesting and shorten
imports.
